Move conv and JSON encoding related logic to new file
